Python Program: Handwritten DIgits Recognition/requirements.txt
Flask
Redis
keras
Pillow
cassandra-driver
tensorflow

